starting modification infos
