--- src/main/java/com/scienceminer/nerd/service/SoftwareInfo.java
@@ -1,37 +1,0 @@
-package com.scienceminer.nerd.service;
-
-public class SoftwareInfo {
-
-
-    String software;
-
-    String version;
-
-    String description;
-
-    public String getSoftware() {
-        return software;
-    }
-
-    public void setSoftware(String software) {
-        this.software = software;
-    }
-
-    public String getVersion() {
-        return version;
-    }
-
-    public void setVersion(String version) {
-        this.version = version;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-
-}
