+++ src/main/resources/service.properties
@@ -1,0 +1,4 @@
+## This file contains only placeholder since there is no other single configuration file
+name=${project.name}
+version=${project.version}
+description=${project.description}
